+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './pokemon';
+export * from './cards';
+export * from './combat';